pyproject.toml
# pyproject.toml for jcook3701.utils
#
# SPDX-FileCopyrightText: Jared Cook
# SPDX-License-Identifier: AGPL-3.0-or-later
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU Affero General Public License as
# published by the Free Software Foundation, either version 3 of the
# License, or (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU Affero General Public License for more details.
#
# You should have received a copy of the GNU Affero General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.
[build-system]
requires = ["setuptools>=80.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jcook3701.utils"
version = "1.0.0"
description = "Ansible utilities for assisting automation processes."
authors = [
    { name="Jared Cook", email="jcook3701+github@gmail.com" }
]
readme = "./README.md"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE.md"]
requires-python = ">=3.11"
dependencies = [
  "CairoSVG>=2.8",
  "pillow>=12.1",
  "PyYAML>=6.0",
  "ruamel.yaml>=0.17.9,<0.18.0",
  "types-Pillow>=10.2",
  "types-PyYAML>=6.0",
  "types-requests>=2.32"
]
keywords = [
  "ansible",
  "galaxy",
  "python3",
  "scripts",
  "automation",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/jcook3701/jcook3701.utils"
Repository = "https://github.com/jcook3701/jcook3701.utils"
Documentation = "https://github.com/jcook3701/jcook3701.utils/docs/jekyll"

[project.optional-dependencies]
dev = [
  "ansible>=12.0",
  "ansible-core>=2.0",
  "ansible-runner>=2.4",
  "galaxy-importer>=0.4",
]
tools = [
  "antsibull-changelog>=0.34",
  "antsichaut>=0.4",
  "bump-my-version>=1.2",
  "black>=25.11",
  "codespell>=2.4",
  "cookiecutter-project-upgrader>=1.3",
  "coverage>=7.12",
  "deptry>=0.24",
  "djlint>=1.36",
  "json-lint-tools>=0.1",
  "jinja2-cli>=0.8",
  "jinja2-time>=0.2",
  "mypy==1.9",
  "nameisok>=0.1",
  "nutri-matic>=0.1",
  "pip-audit>=2.10",
  "pre-commit>=4.5",
  "pytest>=9.0",
  "pytest-ansible>=1.3",
  "pytest-cov>=7.0",
  "pytest-custom-exit-code>=0.3.0",
  "ruff>=0.14.0",
  "tomllint>=0.3",
  "yamllint>=1.0",
  "yamlfixer-opt-nc>=0.9",
]
docs = [
  "sphinx>=8.0",
  "sphinx-rtd-theme>=3.0",
  "sphinx-markdown-builder>=0.6",
  "sphinx-autodoc-typehints>=3.5",
  "ansible-autodoc>=0.5",
  "myst-parser>=0.18",
]
# ====================================
# Setuptools Configuration:
# ====================================
[tool.setuptools.packages.find]
where = [
  "plugins",
]
include = [
  "action*",
  "filter*",
  "inventory*",
  "lookup*",
  "module_utils*",
  "modules*",
]
# ====================================
# Black Configuration: (Formatting)
# ====================================
[tool.black]
line-length = 88          # Maximum line length (default is 88)
target-version = ["py311"] # Target Python version(s)
skip-string-normalization = false  # Whether to skip normalizing string quotes
include = '\.pyi?$'       # Which files to include (regex)
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
# ====================================
# Version Bumping (bump-my-version):
# ====================================
[tool.bumpversion]
current_version = "1.0.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
tag = false
commit = false

[[tool.bumpversion.files]]
filename = "galaxy.yml"
[[tool.bumpversion.files]]
filename = "Makefile"
[[tool.bumpversion.files]]
filename = "./docs/cookiecutter_input.json"
[[tool.bumpversion.files]]
filename = "./docs/jekyll/Makefile"
[[tool.bumpversion.files]]
filename = "./docs/jekyll/_config.yml"
[[tool.bumpversion.files]]
filename = "./docs/sphinx/conf.py"
# ====================================
# Codespell Configuration:
# ====================================
[tool.codespell]
# files/patterns to ignore
skip = [
  "./changelogs/*",
  "cspell.json",
  "./docs/jekyll/_site",
  "*.svg",
  "*.css",
  "*.html",
  "*.lock",
]

# words to ignore globally
ignore-words-list = ["crate", "ba"]

# extra options
check-filenames = true
quiet-level = 2
# ====================================
# Deptry Configuration:
# ====================================
[tool.deptry]
ignore_notebooks = true
exclude = [
    "tests/*",
    "docs/*",
]
# ====================================
# Ruff Configuration: (Linting)
# ====================================
[tool.ruff]
# Target Python version
target-version = "py311"
# Directories to lint and format
src = [
  "plugins",
  "roles",
  "tests",
]
# Exclude these paths
exclude = [
  "playbooks",
  ".venv",
  "build",
  "dist",
  "__pycache__",
  "docs",
  "*.egg-info",
  ".git",
  ".github",
]
# Maximum line length
line-length = 120
# ====================================
# Ruff Lint Configuration:
# ====================================
[tool.ruff.lint]
# Enable linting rules
select = [
  "W",  # PyCodeStyle warnings
  "E",  # pycodestyle errors
  "F",  # pyflakes
  "I",  # isort (import sorting)
  "N",  # pep8-naming
  "UP", # pyupgrade
  "B",  # flake8-bugbear
  "A",  # flake8-builtins
  "C4", # comprehensions
  "TID", # tidy imports
  "FA", # future annotations
  "PL", # pylint-style checks
  "RUF", # ruff-specific
]
# Ignore specific rules
ignore = [
  "E501",  # line too long (handled by formatter)
  "PLR2004",  # magic value comparison
  "B008",  # function call in default argument
]
# ====================================
# Import Sorting Configuration: (Python helper scripts)
# ====================================
[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = [
  "plugins",
]
# ====================================
# Formatter Configuration:
# Ruff also formats YAML, TOML, and Markdown
# ====================================
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
skip-magic-trailing-comma = false
docstring-code-format = true
# ====================================
# Mypy (type-checker) Configuration:
# ====================================
[tool.mypy]
python_version = "3.11"
mypy_path = [
  "plugins",
]
namespace_packages = true
strict = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
# ====================================
# Pytest Configuration:
# ====================================
[tool.pytest.ini_options]
minversion = "7.0"
addopts = ["-v", "--maxfail=1", "-W always", "-rw", "--strict-markers"]
testpaths = ["tests"]
pythonpath = ["plugins"]
